Splitting Internal and External Enclave Methods
